Prevent account creation race condition
Added additional connection flag to ensure a fully formed connection before allowing operations that depend on a fully formed connection, such as account creation.  If the async connection operation has not completed and an account create command is attempted, an account will be created using incomplete network information, causing an invalid WIF to be generated.